inx/outx fixed

+++ kernel/include/utils.h
@@ -11,10 +11,10 @@ const void *memchr(const void *t, u32 v, u32 n);
 void *memset(void *dest, u32 v, u32 n);
 
 /* port actions */
-void outb(u8 p, u8 d);
-u8 inb(u8 p);
+void outb(u16 p, u8 d);
+u8 inb(u16 p);
 
-void outw(u8 p, u16 d);
-u16 inw(u8 p);
+void outw(u16 p, u16 d);
+u16 inw(u16 p);
 
 #endif /* UTILS_H_ */
